\name{customtrees-package}
\alias{customtrees-package}
\alias{customtrees}
\docType{package}
\title{
\packageTitle{customtrees}
}
\description{
\packageDescription{customtrees}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{customtrees}
\packageIndices{customtrees}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{customtrees}

Maintainer: \packageMaintainer{customtrees}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
